Integrated marching cubes + cleanup + organized
Implemented basic encoding of point cloud from a single frame into voxel grid and used marching cubes from previous project to render it. So far, for simplicity color isn't being transferred to the voxel field and density is treated as binary (either a voxel cell is in or out of the volume). The next step will be to implement a more accurate density/color transfer method between the point cloud and voxel field using kernel functions similar to SPH.