Dynamically sized transform feedback buffer
Marching cubes can now operate on voxel grids as big as 500x500x500 if not bigger (Note: I set TDR delay to 10 seconds instead of 2). Saved a ton memory by implementing on demand allocation for the triangle primitives generated by the marching cubes geometry shader by splitting it into 2 passes (the first to count how much space needs to be allocated, the second to generate the triangles as before). In doing so, I implemented a parallel array sum in a compute shader which could easily be generalized to a parallel reduce implementation.